Deduplicate gerrit items when merging
When defining account with multiple aliases it happens that gerrit reports the
same item for both aliases. This change fixes this behaviour by using set()
instead of list and thus deduplicating items. To make this work we define __eq__
and __hash__ functions in Change Gerrit class.

This could be useful for other plugins as well - when multiple people work on
the same item it is (presumably) enough to report it once and not once for each
person when merging.