Add searching for work in progress Gerrit changes
Add support for searching CLs marked as "Work in progress". These are
CLs for which the code review hasn't started yet. The submitted CLs
report is changed to omit these WIP CLs. Additionally, remove support
for drafts ('is:draft') which isn't available in Gerrit anymore and
probably had the same purpose.