gerrit: add a new configuration key to disable wip queries
Not all gerrit servers support the is:wip search. Trying to search
for it on such servers lead to a 400 error.
A new flag can be added to each server to disable wip-based searches.
The default behavior is unchanged, as hopefully more and more
server will support this feature.
The code provides the groundwork to support more of such feature flags.

Fixes #193.