Do not break sorting on merge
My original code for deduplicating in db503e15 was naive and causes another
minor problem when ordering of reported items get reshuffled by set().

This change makes it a little more conservative and when merging we iterate over
list and only append when merging. This way ordering is preserved better and not
randomized.