Add support for shortened 1MT image names
This commit allows images to be specified as with 1minutetip
script. Shortened names, such as fedora, f31, fc31, can now
be used. The same applies to RHEL and CentOS names.

Image name validity is also checked to give a meaningful error
message when the given image is not a valid 1minutetip image
instead of throwing an error due to failed provision.

Signed-off-by: František Nečas <fifinecas@seznam.cz>