Support short options for selecting tests & plans
Selecting tests or plans (for example to be executed or listed) is
a very common use case. Let's add short selection options to the
following commands:

    tmt run test
    tmt run plan
    tmt test ls
    tmt plan ls
    tmt test show
    tmt plan show
    tmt test lint
    tmt plan lint

We keep long options only for story commands because of the
collision with '-c' which is already used for '--covered':

    tmt story ls
    tmt story show
    tmt story coverage

Also keep long options only for export because of possible future
collision with '-f' for '--force' when overwriting existing files.

    tmt story export
    tmt test export