Always run Login plugin even if step is done
This allows connecting to a machine provisioned in the last run.
Previously, the step would be skipped even though the Login plugin was
newly specified on the terminal. This enables the following workflow:

tmt run provision -h container         to provision a machine
tmt run -l login                       to connect to the machine

Signed-off-by: František Nečas <fifinecas@seznam.cz>