Extra check for required packages when using yum
Because of BZ#1920176 yum does not report a problem when both
available and missing packages are provided in a single command.
Thus we add an extra check after the installation is complete.

Also substitue `dnf/yum --cacheonly` with `rpm -q --whatprovides`
in order to speed up the package preparation and prevent asking
users for sudo password even when all required packages are
already installed on the localhost.

Resolves #516.