Enough of dreaming, let's go to the forest! :)
It seems that most of our outlined dreams have already become
true. It's time to forget about dreams and put together a nice
guide which would help new users to walk through the features
which are eagerly waiting for them. What about a couple of
bite-size kind-of stories which will demonstrate essential
features on tangible examples?

In addition to writing The First Steps section and removal of the
dreams pages I've moved the only unfulfilled dream to the stories
and fixed a trivial typo in the README.