Ensure rsync is installed on the guest if needed
To sync workdir `rsync` is needed on both host & guest.
Move the `rsync` requires to the main `tmt` package.
Implement a `requires()` method for the provision step.
Update `local` and `podman` plugins accordingly.
Allow to `login` even if `rsync` is not installed.
Adjust `Guest.execute()` to accept `cwd=None` as well.

Resolves #437 and fixes #558 as well.