Implement shared worktree for the tests
Implement a shared tree directory for running the tests. This is
used for prepare and finish steps when running commands and solves
the problem with weird working directory when running the tests.

Data is synced in via rsync when creating the plan, we omit by
default the .git directory from syncing.

Added two tests to verify that everything works as expected.

Resolves #423

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>