Make run --force behave more expectably
Previously, tmt run -f behaved differently from the other common flags
(such as verbose). Not only was it inehrited into children (steps), but
it also removed the run workdir which could lead to surprising the user
if they only wanted to rerun the selected steps, without having to
specify -f to all of them manually.

Make --force behaviour more in line with the other flags and add a new
option --scratch that follows the behaviour of former --force (remove
the whole workdir).

Signed-off-by: František Nečas <fnecas@redhat.com>