Add support for reboot in a reused provision
Previously, the reboot implementations would be only synced the first
time the plan was run. The token would be however updated in the second
run and hence reboot would not be correctly detected. Backup the
previous {rstrnt, rhts, tmt}-reboot implementations and then override
them always with the new one.

Signed-off-by: František Nečas <fnecas@redhat.com>