Simplify plugin keys handling in wake() and show()
Simplify and unify plugin keys handling to prevent wake() and
show() code duplication in individual plugins. Plugin-specific
keys are stored in `_keys` and keys which are common for all
implementations of a step in `_common_keys`. Both wake() and
show() now default to process these two lists and use the `keys`
parameter to allow processing selected keys only.

This also fixes html plugin which was missing the `wake()` method
and thus the command line option `--open` had no effect.