Switch from fiscal year to natural year
Rather than a natural year, operands for years were using March as start date.
I think this is related to when Red Hat was an independent company and the
fiscal year used to start in March. Now that Red Hat alings with IBM's fiscal
year I think it's not needed anymore.