Use `where` instead of `on` in the multihost spec
In the multihost testing specification rename the on: key to where:.
`on` parses as boolean true in YAML 1.1, so it can lead to potentially
unexpected results when using any YAML 1.1 libraries and tools (linters,
editors, ...).

Using `on` also means that one has to enforce parsing it as string in
the implementation (for example, by forcing YAML 1.2). It will simplify
the code if one does not have to do that, especially when a YAML 1.2
compatible library is not present (such as in EPEL 8).

Another options considered and rejected: `host` (is confusing because it
is similar, but not identical, to Ansible's `hosts`), `hosts` (TMT
usually uses singular, not plural forms, also the syntax is not
identical to Ansible, despite the similar meaning), `onhost` (TMT
typically uses one-word names).

`where` has the advantage of being consistent with the already existing
`how`.