Add support for reboot in interactive mode
The output in interactive mode is forwarded to tmt's output rather than
to a log, hence the previous approach of detecting reboot based on
specific string in the test output was not sufficient. Switch to a
different mechanism to overcome this -- a new file signalling a reboot
request is created by the test which the executor can detect and act
upon after the file is pulled from the guest.

Signed-off-by: František Nečas <fnecas@redhat.com>