Reduce the number of execute calls for reboot
Previously, multiple guest.execute() calls were used in order to setup
and teardown the reboot script. Such calls may however be costly due to
establishing of an ssh connection in the case of provision -h connect.
Reduce the number of calls by pushing a setup and teardown script to the
guest and then executing just these two scripts to handle the setup and
teardown logic.

Signed-off-by: František Nečas <fnecas@redhat.com>