Force order of clean operations
Running tmt clean runs guests with enabled command chaining could lead
to a nasty error where runs are removed before guests and hence then
there is no information about the guests and the user would have to
clear the VMs/containers manually which is not desired.

Force guests to be always cleaned before runs to prevent this possible
user error.

Signed-off-by: František Nečas <fnecas@redhat.com>