Split `Guest` class to separate SSH-capable guests
The base class for all guests was expecting some parameters that were
useful only in the context of SSH-capable, "connectable" guest. For
example Podman-backed guests have no use for `guest` or `password` keys.
To express this hierarchy better, and to open it for more strict types
involved in various `data` packages passed between methods, this patch
splits `Guest` class into a base one and a new one, representing a guest
with SSH capability.

Guest implemenations are now based on the new `GuestSsh` class, with the
exception of `GuestContainer` and `GuestLocal` which are derived from
`Guest` since it does not use SSH.