Avoid re-using image/instance for different values by testcloud plugin
`image` is a key given by user, and it's internally overshadowed by the
attribute of the same name being used to hold a testcloud image
instance. Together with hard to spot difference between `image` and
`image_url` and `instance` being saved as `instance_name`, this created
a bit of a entangled web complicating addition of type annotations.

Patch tries to clear things up a bit by adding attributes dedicated to
holding runtime testcloud objects, never to be intermixed with keys. And
fixes dicrepancy between saved keys and their names in `data` container.