Avoid Library url conflict if repo doesn't exist
Upstream beakerlib library url is guessed by tmt and might not
exist.  In that case there can't be conflict with already fetched
repo of same name.

Doing this only for 'rpm' translated strings.  If user specifies
url in metadata directly conflicting url erorr is raised without
fetching the repo to not silence issues.

Fixes: #1220