Convert FmfIdType from TypedDict to a dataclass
While TypedDict is imposed to a dictionary with no other effect, a
dataclass can serve as a type on its own while providing methods related
to the type. Patch converts fmf id type to a dataclass to both carry the
values and to tie "validate fmf id" function to the given id stored in
the container. Instead of a dictionary and a function, we now have a fmf
id "instance" with relevant methods directly connected.