Use decorator to register plugin methods
Instead of listing Method instances, and dedicated `_methods` list in each
plugin class, patch adds a class decorator for this task. It's a different
approach, replacing `_methods` and `PluginIndex` metaclass, but it has some
advantages, the most visible being less characters to write:

With regard to #1373, the necessity of annotating each `_methods`
instance would become even more painful as soon as `BasePlugin` would
become a generic type. With the decorator, the list is not needed at
all.

A few more changes included in the decorator to make our lives easier:
the default value for `doc` can be extracted, the default value for `order`
can be found in the code and declared explicitly, and both are used in the
common manner of keyword arguments. With the metaclass, extracting
docstring and default values would be possible too, but would have to be
done as explicit assignments in the code of metaclass, which makes them
less visible.

For the compatibility with out-of-tree plugins, the metaclass and
`_methods` list are still supported, but are deprecated and will emit a
warning.