Inject logger object to base node classes inheritance
Instead of relying on `self.{info,warn,debug}`, a logger object is
injected to all `__init__()` classes involved. This logger might be
`self`, which is actually the value used by `Test`, `Plan` and `Story`
classes use, but the advantage is the explicitness: mixin class - or
classes, in the future - do not need to rely on `self.warn` be defined.
An object providing logging would be clearly given.