Allow raising an exception on validation errors
The original implementation of schema-based validation only reported the
issues, and moved on. This has been preserved, but I discovered there
are tests that focus on failed validation which, once the Python-based
implementation in `tmt.base` is removed, would began to fail.

To accomodate these tests and possible future enforced validation,
validation mixing gains a flag to enable it to raise an exception
instead of a mere warning. The default value of the flag is `False`, so
no change in the behavior for uninformed user.