Teach schema validation tests to peek into other trees
There are some "hidden" trees in tmt's repo, trees that are tailored for
various tests. These are not covered by schema-validation unit test,
this patch adds necessary code to enable this coverage.

It's left disabled, though, because plenty of theese trees are malformed
on purpose, and therefore cause a test failure. Before enabling the test
in general, we need a way how to mark a tree as "broken, nothing to see
here, move on" *for this test only* - we do not want `tmt lint` to skip
it, just the unit test.