Add type annotations to `base.py`
And many other files, since base.py is one of the core pieces of tmt,
and many other modules depend on it or are used by base.py code. This
means mypy is allowed to look around, and reports also conflicts between
existing annotations and new ones that must be amended.