Reorder step and their base plugin classes
(Part of a patchset focusing on a step data conversion to dataclasses.)

There is no functional change, the patch switches order of step class
and its respective base plugin class for each step. The goal is to reach
a state where each step can offer a link to base class of its plugins,
which will become useful in the step data conversion patch.