Enhance ClickOptionDecorator type to announce identity
Current type works, but is not perfect - namely, it lacks any notion of
identity between argument and return value of `click.option` decorators.
This makes it incompatible with how `mypy` sees `click.option`, and
forces us to use a suboptimal code when overriding `options()` method.
We can re-use bits of `click.decorators` code to improve the situation,
and children's `options()` becomes more straightforward.