Ignore plan -n when searching for upgrade path
Previously, options from the plan command on the terminal would also be
propagated to the remote upgrade repo. This is not desired, the plan
should only filter in the local repository, since otherwise it could
cause incorrect upgrade path to be selected (or an error to be thrown
due to multiple upgrade paths being matched).

Signed-off-by: František Nečas <fnecas@redhat.com>