Fix variable redefinition in `discover/shell.py`
`test` variable is assigned two different values, of two different
types, `fmf.Tree` and then a list of tests. The second assignment is
used just to propagate the value to `self._tests`, which we can do right
away.

mypy should have reported this, but is confused because of `fmf.Tree`
being `Any`.