Allow all plugins to fetch secrets from files
Some plugins like the Jira one allowed to store their token used for
authentication within a file an have you point to the file using the
`token_file` config option. I liked that idea because it makes did
config files much more shareable amongst team members.

That's why I've added the token-in-file mechanism to all plugins. The
google plugin also stores a `client_it` and `client_secret` that you can
also put into a file now by specifying `client_id_file` and
`client_secret_file`.

I've added tests for the `get_token` function and for the github plugin
to utilize this function when running in a Github action because then
the environment variable `GITHUB_TOKEN` is set and can be tested when
putting it in a token file only for it to be consumed then.