Capture uncaught exceptions when using testcloud
There are two exceptions when using testcloud. The first one is
"KeyError" when downloading remote image, the second one is "TypeError"
when creating a testcloud instance but the instance name is "None".
For the first exception, just capture it. But for the second one, we
need to make sure the instance name is set.

Signed-off-by: Vector Li <huanli@redhat.com>