Include the new web link in verbose `show` mode
Show the clickable web link when showing tests, plans and stories
in verbose mode. Also include fmf-id in the story and plan output.
Allows to quickly access the source code from the web browser.

Also fix the test so that it works fine also for pull requests
from forked repositories.