Produce fixed `phabricator` statistics
This brings much better statistics to the phabricator plugin that can be
individually queried with these command line options:

    --<section>-differentials-created
    --<section>-differentials-closed
    --<section>-differentials-accepted
    --<section>-differentials-changes-requested
    --<section>-differentials-commented

Each section returns ordered results and these results can naturally
overlap.

We find the Differentials that might be relevant for the given user and
then fetch the transaction events for each Differential to better
categorize it and put it in an appropriate bucket (e.g. closed,
accepted, etc.). If in the future there's interest in more statistics,
this mechanism allows for more categorization.

Differentials are stored and passed around in sets to avoid any
duplication. The statistics based upon these sets are converted into
sorted lists to have a predictable output.

Verbosity output is now managed with a class variable to avoid storing
the verbosity in each Differential.

The tests no longer check for the title of a found Differential but only
for it's ID. This way we compensate for a Differential's title that
may change over time.

There's now a general mechanism to fetch paged results from phabricator
if the endpoint supports it.

Transactions events for each Differential are fetched in parallel to
speed up the I/O.

Fix: https://github.com/psss/did/issues/288