Additional filter for commented issues and PRs
GH search API unable to identify if the issues and PRs were updated by
the user's comment. To make it reliable, one needs to search the
comments for the entity after the `since` date, sort them by created_at
field, and check if the comments in the given interval are created by
the user.