public_inbox: Make __fetch_thread_root() return optional value
Commit af2b8ac97122 ("Fix pylint R1710") made __fetch_thread_root()
return the original message argument if no thread root could be found.

However, and even though it wasn't really handled yet, the intent was
always to return None in such a case.

Restore the previous behaviour, but in an explicit way this time, and
we'll handle it in the callers in the next commit.

Signed-off-by: Maxime Ripard <mripard@kernel.org>