public_inbox: Handle missing parent is __get_thread_root()
A message can have an In-Reply-To header but the message it points to
might not be available for some reason.

In such a case, the Message.parent_id() method will return a proper
value, but __fetch_thread_root() on that message might not and would
return None.

Let's handle that case by returning the latest message we found as the
root message.

Signed-off-by: Maxime Ripard <mripard@kernel.org>