Avoid fetching thousands of pages from GitLab API
When you comment on an issue in a project that has thousands of issues,
or operate on a merge request in a project with thousands of MRs, did
attempts to fetch them all to identify the ID of the one you interacted
with.

This is OK for projects that are small in size, but becomes a huge issue
for large projects such as https://gitlab.com/gitlab-org/gitlab, which
has 218k issues.

Check for the number of pages that would be fetched and avoid loading
them all if there are more than 20 pages. This will lead to issue and
merge request IDs missing from the report, but it will allow the report
to finish within your lifetime.

Signed-off-by: Clemens Lang <cllang@redhat.com>