Add stats for updated jira tickets
Previously a ticket in jira was considered updated only if commented.
Now commented tickets are in their own stats and tickets are considered
updated if any type of update is done on them according to their
changelog.
In order to limit the amount of data being processed, `project` now
support a list of projects instead of a single one and it's required for
collecting updates.

Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>