Use custom `user-agent` in the `public_inbox` plugin (#392)
The lore.kernel.org mailing-list archive recently started using the
Anubis proxy to filter out AI scrap-bots.

Unfortunately, it seems to be somewhat based on the User-Agent, and the
requests default one will get rejected. This means that the public_inbox
will now be completely broken.

This can be bypassed by setting a custom User-Agent though, which
restores the expected behaviour, so let's do so.

Signed-off-by: Maxime Ripard <mripard@kernel.org>