Fix PageModified stat in Confluence plugin
Extracted `fetch_ratelimited_url` from `search` in order to avoid
duplicated code and used it in the new method `get_page_versions`.

Used this new method to ensure the page modified in the given time range
have been really modified by the user and not just created by the user
in the past but modified by someone else.

Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>